db management split
